+++ src/index.ts
@@ -7,6 +7,7 @@
 
 export * from "./components/primitives";
 export * from "./components/select";
+export * from "./components/popover";
 export * from "./components/theme";
 export * from "./components/alerts";
 export * from "./components/navbar";